Fixed namespace pollution issues. Changed namespace WeatherAnalysis to TextGen.
using namespace X is not standard practice in C++ headers, it leads to
namespace pollution issues since the user cannot control what namespaces are
visible. Fixed by removing all "using namespace" commands from the
headers and then fixing the warnings that followed by using the appropriate
namespaces.

Also, it is unnecessary for a single library to provide multiple namespaces.
It complicates code by forcing namespace usage in the API. Replaced namespace
WeatherAnalysis with TextGen, which is now the top namespace in the entire
library.