wind_overview- and temperature_max36_hours-story fixes
- No minus('-') sign with sub-zero temperatures when 'pakkanen'-phrase is used: 'Pakkasta vajaat -5 astetta' -> 'Pakkasta vajaat 5 astetta'
- 'Vajaat nolla astetta'-phrase corrected to form 'VÃ¤hÃ¤n nollan alapuolella'
- new phrases to database ('vÃ¤hÃ¤n nollan alapuolella')
- Highcharts graphs (in wind_overview story) improved for better readibility
- new configuration parameters:
* wind_speed_top_coverage (default value 98.0): This parameter tells how many percent the reported
top wind speed must cover of the total area.
For example if top wind on the area is 11 m/s, but it is predicted only on 1.9 % of total area, the reported top wind is 10 m/s.
* wind_direction_min_speed (oletusarvo 6.5): This is limiting value of wind speed for wind direction when
standard deviation of wind direction is big (> 45.0).
If standard deviation of wind direction is big and top wind speed is smaller or equal
than 'wind_direction_min_speed' wind direction is reported as varying.
- code cleaned and refactored